backend/pyproject.toml
[tool.poetry]
name = "moon"
version = "0.1.0"
description = ""
authors = ["tangkikodo <allmonday@126.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "0.100"
pydantic-resolve = "^1.6.3"
httpx = "^0.24.1"
python-dateutil = "^2.8.2"
sqlalchemy = "^2.0.19"
pymysql = "^1.1.0"
asyncmy = "^0.2.8"
alembic = "^1.11.1"
fastapi-users = {extras = ["oauth", "sqlalchemy"], version = "^12.1.0"}
uvicorn = {extras = ["standard"], version = "^0.23.2"}


[tool.poetry.group.dev.dependencies]
autopep8 = "^2.0.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
